Use MUnit in ByteToObjectStageSuite
